update : config file from the Internet
