+++ src/main/java/com/github/sakakiaruka/cutomcrafter/customcrafter/CustomCrafter.java
@@ -2,8 +2,8 @@ package com.github.sakakiaruka.cutomcrafter.customcrafter;
 
 import com.github.sakakiaruka.cutomcrafter.customcrafter.commands.CheckObject;
 import com.github.sakakiaruka.cutomcrafter.customcrafter.listeners.Listeners;
+
 import com.github.sakakiaruka.cutomcrafter.customcrafter.some.SettingsLoad;
-import org.bukkit.plugin.Plugin;
 import org.bukkit.plugin.java.JavaPlugin;
 
 public final class CustomCrafter extends JavaPlugin {
