+++ src/main/resources/config.yml
@@ -11,127 +11,145 @@
 # Do not name that the length less than 3.
 # Have to use sentences.
 ######
-base: gold_block
-
-result-items:
-  - first
-  - second
-  - emerald
-  - big-torch
-  - empty-bucket
-result-item:
-  first: # test section
-    material: iron_block #need
-    amount: 2 #need -> 0 < amount < common sense
-
-  second:
-    material: gold_ingot
-    amount: 10
-
-    lore:
-      - line 1
-      - line 2
-    enchants:
-      - mending,10
-      - luck,2
-    flags:
-      - hide_dye
-
-  emerald:
-    material: emerald
-    amount: 3
-    display-name: "This item is an emerald."
-
-  big-torch:
-    material: torch
-    amount: 10
-    display-name: "BIG_TORCH"
 
 
+#base: gold_block
+#
+#result-items:
+#  - first
+#  - second
+#  - emerald
+#  - big-torch
+#  - empty-bucket
+#result-item:
+#  first: # test section
+#    material: iron_block #need
+#    amount: 2 #need -> 0 < amount < common sense
+#
+#  second:
+#    material: gold_ingot
+#    amount: 10
+#
+#    lore:
+#      - line 1
+#      - line 2
+#    enchants:
+#      - mending,10
+#      - luck,2
+#    flags:
+#      - hide_dye
+#
+#  emerald:
+#    material: emerald
+#    amount: 3
+#    display-name: "This item is an emerald."
+#
+#  big-torch:
+#    material: torch
+#    amount: 10
+#    display-name: "BIG_TORCH"
+#
+#
+#
+#material-category:
+#  - test
+#  - logs
+#
+#material-category-contents:
+#  test:
+#    - diamond_ore
+#    - diamond
+#
+#  logs:
+#    - oak_log
+#    - spruce_log
+#    - birch_log
+#    - jungle_log
+#    - dark_oak_log
+#    - acacia_log
+#    - crimson_stem
+#    - warped_stem
+#
+#
+#recipe-materials-list:
+#  - first
+#  - second
+#  - third
+#  - logs
+#  - enchanted
+#
+#recipe-materials:
+#  first:
+#    material: coal # need
+#    amount: 1 # need (default:1 -> when you set this over 1, rarely occur errors.
+#
+#  second:
+#    material: stone
+#    amount: 1
+#
+#  third:
+#    mixed-material: test
+#    amount: 1
+#
+#  logs:
+#    mixed-material: logs
+#    amount: 1
+#
+#  enchanted:
+#    name: enchanted
+#    material: fishing_rod
+#    amount: 1
+#    enchants:
+#      - lure,1,OnlyEnchant
 
-material-category:
-  - test
-  - logs
+#recipe-list:
+#  - stone-to-iron
+#  - iron-to-gold
+#  - diamond-to-emerald
+#  - big-torch
+#
+#recipes:
+#  stone-to-iron:
+#    size: 3 # 3 <= size <= 6 (int)
+#    result: first
+#    recipe:
+#      - second,second,second
+#      - second,null,second
+#      - second,second,second
+#
+#  iron-to-gold:
+#    size: 4
+#    result: second
+#    recipe:
+#      - first,first,null,null
+#      - null,null,null,first
+#      - first,first,null,null
+#      - null,null,first,first
+#
+#  diamond-to-emerald:
+#    size: 3
+#    result: emerald
+#    recipe:
+#      - null,third,null
+#      - null,null,third
+#      - third,third,third
+#
+#  big-torch:
+#    size: 3
+#    result : big-torch
+#    recipe:
+#      - first,first,first
+#      - logs,logs,logs
+#      - logs,logs,logs
+#
 
-material-category-contents:
-  test:
-    - diamond_ore
-    - diamond
-
-  logs:
-    - oak_log
-    - spruce_log
-    - birch_log
-    - jungle_log
-    - dark_oak_log
-    - acacia_log
-    - crimson_stem
-    - warped_stem
-
-
-recipe-materials-list:
-  - first
-  - second
-  - third
-  - logs
-
-recipe-materials:
-  first:
-    material: coal # need
-    amount: 1 # need (default:1 -> when you set this over 1, rarely occur errors.
-
-  second:
-    material: stone
-    amount: 1
-
-  third:
-    mixed-material: test
-    amount: 1
-
-  logs:
-    mixed-material: logs
-    amount: 1
-
-recipe-list:
-  - stone-to-iron
-  - iron-to-gold
-  - diamond-to-emerald
-  - big-torch
-
-recipes:
-  stone-to-iron:
-    size: 3 # 3 <= size <= 6 (int)
-    result: first
-    recipe:
-      - second,second,second
-      - second,null,second
-      - second,second,second
-
-  iron-to-gold:
-    size: 4
-    result: second
-    recipe:
-      - first,first,null,null
-      - null,null,null,first
-      - first,first,null,null
-      - null,null,first,first
-
-  diamond-to-emerald:
-    size: 3
-    result: emerald
-    recipe:
-      - null,third,null
-      - null,null,third
-      - third,third,third
-
-  big-torch:
-    size: 3
-    result : big-torch
-    recipe:
-      - first,first,first
-      - logs,logs,logs
-      - logs,logs,logs
-
-
-wget:
-  - "https://github."
+baseBlock: "plugins/Custom_Crafter/baseBlock"
+resultItems: "plugins/Custom_Crafter/resultItems"
+materialCategory: "plugins/Custom_Crafter/materialCategory"
+recipeMaterials: "plugins/Custom_Crafter/recipeMaterials"
+recipes: "plugins/Custom_Crafter/recipes"
+download:
+  - "wget https://github.com/Sakaki-Aruka/custom-crafter-config/raw/master/baseBlock.yml -P plugins/Custom_Crafter/baseBlock/"
+  - "wget https://github.com/Sakaki-Aruka/custom-crafter-config/raw/master/recipe-material-second.yml -P plugins/Custom_Crafter/recipeMaterials/"
+  - "wget https://github.com/Sakaki-Aruka/custom-crafter-config/raw/master/recipes-stone-to-iron.yml -P plugins/Custom_Crafter/recipes/"
+  - "wget https://github.com/Sakaki-Aruka/custom-crafter-config/raw/master/result-item-first.yml -P plugins/Custom_Crafter/resultItems/"
